requirements.txt
# GUI Framework
PyQt6>=6.10.0

# Audio Processing
pyaudio==0.2.14
numpy==1.26.3

# APIs
openai>=2.7.0
anthropic>=0.72.0

# Google Drive
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.114.0

# Utilities
python-dateutil>=2.8.2

# Mac App Building
py2app>=0.28.7
